requirements.txt
torch==2.9.1
torchvision==0.24.1
numpy==2.4.0
pillow==12.0.0
opencv-python
scikit-learn
matplotlib
omegaconf
jupyter
